Added arduino communication functionality+problems persisits
This sends and receives data finally.
But if a device is reconnected then the stream of data stored in output stream of arduino will accumulate in arduino and be send in a single thrust to android. This is old and useless data. while(BTserial.available == 0) helps to avaoid this for the *first time*, but not on reconnection.